feat(transfer-uri): add package with bip-321 and erc681 parsers
