+++ packages/suite/tsconfig.json
@@ -135,6 +135,9 @@
             "path": "../../suite-common/transaction-search"
         },
         {
+            "path": "../../suite-common/transfer-uri"
+        },
+        {
             "path": "../../suite-common/tx-simulation"
         },
         {
