+++ suite-common/suite-utils/src/__tests__/protocol.test.ts
@@ -1,20 +1,12 @@
 import { type Protocol } from '@suite-common/suite-constants';
 
 import * as fixtures from '../__fixtures__/protocol';
-import { getNetworkSymbolForProtocol, parseErc681TransferUri } from '../protocol';
+import { getNetworkSymbolForProtocol } from '../protocol';
 
 describe('getNetworkSymbolForProtocol', () => {
     fixtures.getNetworkSymbolForProtocol.forEach(f => {
         it(f.description, () => {
             expect(getNetworkSymbolForProtocol(f.scheme as Protocol)).toEqual(f.result);
-        });
-    });
-});
-
-describe('parseErc681TransferUri', () => {
-    fixtures.parseErc681TransferUri.forEach(f => {
-        it(f.description, () => {
-            expect(parseErc681TransferUri(f.uri)).toEqual(f.result);
         });
     });
 });
