+++ suite-common/transfer-uri/package.json
@@ -1,0 +1,20 @@
+{
+    "name": "@suite-common/transfer-uri",
+    "version": "1.0.0",
+    "private": true,
+    "license": "See LICENSE.md in repo root",
+    "sideEffects": false,
+    "main": "src/index",
+    "scripts": {
+        "depcheck": "yarn g:depcheck",
+        "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
+        "test:unit": "yarn g:jest",
+        "type-check": "yarn g:tsc --build"
+    },
+    "dependencies": {
+        "@suite-common/suite-constants": "workspace:*",
+        "@suite-common/suite-utils": "workspace:*",
+        "@suite-common/wallet-config": "workspace:*",
+        "@trezor/type-utils": "workspace:*"
+    }
+}
