+++ suite-common/transfer-uri/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "compilerOptions": { "outDir": "libDev" },
+    "references": [
+        { "path": "../suite-constants" },
+        { "path": "../suite-utils" },
+        { "path": "../wallet-config" },
+        { "path": "../../packages/type-utils" }
+    ]
+}
