+++ suite-native/module-send/package.json
@@ -22,9 +22,9 @@
         "@suite-common/device": "workspace:*",
         "@suite-common/formatters": "workspace:*",
         "@suite-common/suite-sync": "workspace:*",
-        "@suite-common/suite-utils": "workspace:*",
         "@suite-common/token-definitions": "workspace:*",
         "@suite-common/transaction-search": "workspace:*",
+        "@suite-common/transfer-uri": "workspace:*",
         "@suite-common/validators": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
         "@suite-common/wallet-constants": "workspace:*",
