+++ suite-native/module-send/tsconfig.json
@@ -14,13 +14,13 @@
             "path": "../../suite-common/suite-sync"
         },
         {
-            "path": "../../suite-common/suite-utils"
-        },
-        {
             "path": "../../suite-common/token-definitions"
         },
         {
             "path": "../../suite-common/transaction-search"
+        },
+        {
+            "path": "../../suite-common/transfer-uri"
         },
         {
             "path": "../../suite-common/validators"
